requirements.txt
# ─── Core computer-vision / data ──────────────────────────────────────────────
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
Pillow==10.3.0
reportlab==4.2.0

# ─── Web / API ────────────────────────────────────────────────────────────────
flask==3.1.0
flask-restx==1.3.0
flask-jwt-extended==4.7.1
flask-limiter==3.9.0
flask-wtf==1.2.2
Werkzeug==3.1.3
gunicorn==23.0.0

# ─── Security ────────────────────────────────────────────────────────────────
bcrypt==4.2.1

# ─── Configuration ────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Firebase (optional — required when USE_FIREBASE=true) ───────────────────
firebase-admin==6.6.0

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-flask==1.3.0

# ─── Excel reports ──────────────────────────────────────────────────────────
openpyxl==3.1.2

# ─── Email alerts ───────────────────────────────────────────────────────────
Flask-Mail==0.9.1

# ─── Database ORM ───────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# ─── CORS support ───────────────────────────────────────────────────────────
flask-cors==4.0.0